MdeModulePkg/PciBusDxe: Don't create resource node if BAR already assigned
While collecting resource nodes, skip BARs that have non-zero addresses
already programmed into them and inform GCD of those resources instead.
We treat these BARs as pre-assigned and do not relocate them.

This needs to be reworked later as it's considered an OVMF-specific
behavior. The PCI spec does not stipulate the initial value of a BAR.

Signed-off-by: Peter Fang <peter.fang@intel.com>