[REVERTME] OvmfPkg: Relocate to 0xFFC00000 for 2MB images
Relocate the base address of the 2MB image from (4GB - 2MB) to (4GB -
4MB).

NOTE:
This is very fragile and has only been tested using IntelTdx images. It
only works with FD_SIZE_2MB. Functions such as
Flat32SearchForSecEntryPoint can be problematic in corner cases, but a
good BIOS image is not expected to hit them.

Signed-off-by: Peter Fang <peter.fang@intel.com>